refactor(F204): reorganize plugin code — limb-plugins package pattern
Follow IM connector plugin-package pattern: platform-specific handler
code lives alongside the generic framework, not in a separate domain.

  domains/weixin-mp/          → domains/limb/limb-plugins/weixin-mp/
  weixin-mp-handlers.ts       → handlers.ts (scoped by directory)
  markdown-to-wx-html.ts      → same, moved into plugin package

Generic framework (PluginLimbAdapter, PluginRestExecutor, PluginTokenManager)
stays in domains/limb/. Plugin declarations (YAML, skills) stay in plugins/.
Per-platform TypeScript now sits next to the framework it extends.

Adding a new platform = plugins/<id>/ (YAML+skills) + limb-plugins/<id>/ (handlers).

[宪宪/claude-opus-4-6🐾]
